Interface agora é única
