+++ src/functions/listSum.function.ts
@@ -1,0 +1,5 @@
+import { IList } from "../interfaces/list.interface";
+
+export default function sumListAmount(list: IList) {
+
+}
